refactored postModal
